Fix jni bugs

+++ android/InspireFaceExample/inspireface/src/main/java/com/insightface/sdk/inspireface/base/SearchTopKResults.java
@@ -3,5 +3,5 @@ package com.insightface.sdk.inspireface.base;
 public class SearchTopKResults {
     public int num;
     public float[] confidence;
-    public int[] ids;
+    public long[] ids;
 }
